angular: Add service method to get global config
When querying the state from the `JsonFormsAngularService` via its `getState` method, the whole state is cloned.
This lead to a heavy performance penalty because the `JsonFormsAbstractControl` queried the state on every change detection cycle to get the global config. Thus, the whole state was cloned for every control on every change.
As the config is only a small part of the state, this problem can be alleviated by adding a getConfig method to the service that only clones the config.

* Extend `JsonFormsAngularService` with a `getConfig` method that returns a clone of the global config
* Use the new method in `JsonFormsAbstractControl`

Signed-off-by: Lucas Koehler <lkoehler@eclipsesource.com>