Fix category hiding in Angular
Previously the category renderer in Angular did not take into account
the visibility of a category and showed them regardless of the status
of the property. Prefilter visible categories before rendering them in
the template.

Fix unit test by moving the setting of the new UI schema with
jsonFormsService after it was reassigned in the component. This way
when the event is triggered on the subscription the new schema is
already present. This is necessary for the new changes and also fixes
the previously commented out part.